add user and login to webapp
