+++ src/tracker/api/templates/operations.html
