+++ src/tracker/api/templates/register.html
@@ -27,7 +27,7 @@
               <script src="/static/script.js">  
             </script>
               <footer>Mam już konto 
-                <a href="#">Zaloguj</a>
+                <a href="/login_form">Zaloguj</a>
               </footer>
           
           </div>
