+++ src/tracker/consts.py
@@ -1,4 +1,4 @@
 class Consts:
 
-    TEMPLATES_PATH = r'C:\Users\aleks\Desktop\Python\AI\budget_manager\src\tracker\app\templates'
-    STATIC_PATH = r'C:\Users\aleks\Desktop\Python\AI\budget_manager\src\tracker\app\static'
+    TEMPLATES_PATH = r'C:\Users\aleks\Desktop\Python\AI\budget_manager\src\tracker\api\templates'
+    STATIC_PATH = r'C:\Users\aleks\Desktop\Python\AI\budget_manager\src\tracker\api\static'
